+++ bridgekeeper/predicate_many_relation_tests.py
@@ -14,10 +14,10 @@ def test_many_relation_to_user():
     user_branch_in_store = ManyRelation(
         'branch_set', 'branch', Branch, Is(lambda u: u.profile.branch))
 
-    assert user_branch_in_store.apply(u1, s1)
-    assert user_branch_in_store.apply(u2, s2)
-    assert not user_branch_in_store.apply(u1, s2)
-    assert not user_branch_in_store.apply(u2, s1)
+    assert user_branch_in_store.check(u1, s1)
+    assert user_branch_in_store.check(u2, s2)
+    assert not user_branch_in_store.check(u1, s2)
+    assert not user_branch_in_store.check(u2, s1)
 
     qs1 = user_branch_in_store.filter(Store.objects.all(), u1)
     qs2 = user_branch_in_store.filter(Store.objects.all(), u2)
@@ -34,4 +34,4 @@ def test_many_relation_never_global():
     user = UserFactory()
     user_branch_in_store = ManyRelation(
         'branch_set', 'branch', Branch, Is(lambda u: u.profile.branch))
-    assert not user_branch_in_store.apply(user)
+    assert not user_branch_in_store.check(user)