+++ docs/changelog.rst
@@ -5,3 +5,4 @@ Changelog
 ===
 
 - Renamed ``bridgekeeper.registry.registry`` to :data:`bridgekeeper.perms`.
+- Renamed ``bridgekeeper.predicates.Predicate.apply()`` to :meth:`~bridgekeeper.predicates.Predicate.check`